Conflicts:
	LDFLAGS
	Makefile
	conformance/interfaces/clock_getres/6-2.c
	conformance/interfaces/clock_gettime/8-2.c
	conformance/interfaces/clock_settime/17-1.c
	conformance/interfaces/clock_settime/17-2.c
	conformance/interfaces/fsync/4-1.c
	conformance/interfaces/mmap/1-1.c
	conformance/interfaces/mmap/1-2.c
	conformance/interfaces/mmap/10-1.c
	conformance/interfaces/mmap/11-1.c
	conformance/interfaces/mmap/11-2.c
	conformance/interfaces/mmap/11-3.c
	conformance/interfaces/mmap/11-4.c
	conformance/interfaces/mmap/11-5.c
	conformance/interfaces/mmap/12-1.c
	conformance/interfaces/mmap/13-1.c
	conformance/interfaces/mmap/14-1.c
	conformance/interfaces/mmap/18-1.c
	conformance/interfaces/mmap/21-1.c
	conformance/interfaces/mmap/24-1.c
	conformance/interfaces/mmap/24-2.c
	conformance/interfaces/mmap/27-1.c
	conformance/interfaces/mmap/28-1.c
	conformance/interfaces/mmap/3-1.c
	conformance/interfaces/mmap/31-1.c
	conformance/interfaces/mmap/5-1.c
	conformance/interfaces/mmap/6-1.c
	conformance/interfaces/mmap/6-2.c
	conformance/interfaces/mmap/6-3.c
	conformance/interfaces/mmap/6-4.c
	conformance/interfaces/mmap/6-5.c
	conformance/interfaces/mmap/6-6.c
	conformance/interfaces/mmap/7-1.c
	conformance/interfaces/mmap/7-2.c
	conformance/interfaces/mmap/7-3.c
	conformance/interfaces/mmap/7-4.c
	conformance/interfaces/mmap/9-1.c
	conformance/interfaces/munmap/1-1.c
	conformance/interfaces/munmap/1-2.c
	conformance/interfaces/munmap/3-1.c
	conformance/interfaces/munmap/4-1.c
	conformance/interfaces/munmap/9-1.c
	conformance/interfaces/pthread_attr_setdetachstate/2-1.c
	conformance/interfaces/pthread_cond_timedwait/2-2.c
	conformance/interfaces/pthread_cond_timedwait/2-3.c
	conformance/interfaces/pthread_detach/1-1.c
	conformance/interfaces/pthread_detach/2-1.c
	conformance/interfaces/pthread_join/speculative/6-1.c
	conformance/interfaces/pthread_key_create/speculative/5-1.c
	conformance/interfaces/pthread_mutexattr_setpshared/3-1.c
	conformance/interfaces/pthread_mutexattr_settype/3-2.c
	conformance/interfaces/pthread_once/3-1.c
	conformance/interfaces/pthread_once/6-1.c
	conformance/interfaces/pthread_sigmask/18-1.c
	conformance/interfaces/sem_close/1-1.c
	conformance/interfaces/sem_close/2-1.c
	conformance/interfaces/sem_close/3-1.c
	conformance/interfaces/sem_getvalue/1-1.c
	conformance/interfaces/sem_getvalue/2-1.c
	conformance/interfaces/sem_getvalue/4-1.c
	conformance/interfaces/sem_getvalue/5-1.c
	conformance/interfaces/sem_init/6-1.c
	conformance/interfaces/sem_open/1-4.c
	conformance/interfaces/sem_open/6-1.c
	conformance/interfaces/sem_post/1-1.c
	conformance/interfaces/sem_post/1-2.c
	conformance/interfaces/sem_post/2-1.c
	conformance/interfaces/sem_post/4-1.c
	conformance/interfaces/sem_post/5-1.c
	conformance/interfaces/sem_post/6-1.c
	conformance/interfaces/sem_post/8-1.c
	conformance/interfaces/sem_timedwait/3-1.c
	conformance/interfaces/sem_timedwait/7-1.c
	conformance/interfaces/sem_unlink/1-1.c
	conformance/interfaces/sem_unlink/2-1.c
	conformance/interfaces/sem_unlink/5-1.c
	conformance/interfaces/sem_wait/1-1.c
	conformance/interfaces/sem_wait/1-2.c
	conformance/interfaces/sem_wait/11-1.c
	conformance/interfaces/sem_wait/12-1.c
	conformance/interfaces/sem_wait/3-1.c
	conformance/interfaces/sem_wait/5-1.c
	conformance/interfaces/sem_wait/7-1.c
	conformance/interfaces/shm_open/23-1.c
	conformance/interfaces/sigqueue/12-1.c
	conformance/interfaces/sigqueue/3-1.c
	conformance/interfaces/sigset/6-1.c
	conformance/interfaces/sigset/7-1.c
	conformance/interfaces/sigset/8-1.c
	conformance/interfaces/sigwait/3-1.c
	conformance/interfaces/sigwait/assertions.xml
	conformance/interfaces/timer_create/10-1.c
	conformance/interfaces/timer_create/11-1.c
	conformance/interfaces/timer_delete/speculative/5-1.c
	conformance/interfaces/timer_delete/speculative/5-2.c
	conformance/interfaces/timer_getoverrun/speculative/6-1.c
	conformance/interfaces/timer_getoverrun/speculative/6-2.c
	conformance/interfaces/timer_getoverrun/speculative/6-3.c
	conformance/interfaces/timer_gettime/speculative/6-1.c
	conformance/interfaces/timer_gettime/speculative/6-2.c
	conformance/interfaces/timer_gettime/speculative/6-3.c
	conformance/interfaces/timer_settime/speculative/12-1.c
	conformance/interfaces/timer_settime/speculative/12-2.c
	conformance/interfaces/timer_settime/speculative/12-3.c

+++ conformance/interfaces/mmap/7-2.c
@@ -46,7 +46,8 @@ int main(void)
 	ssize_t size = 1024;
 	int fd;
 
-	snprintf(tmpfname, sizeof(tmpfname), "/tmp/pts_mmap_7_2_%d", getpid());
+	snprintf(tmpfname, sizeof(tmpfname), "/tmp/pts_mmap_7_2_%ld",
+		(long)getpid());
 	unlink(tmpfname);
 	fd = open(tmpfname, O_CREAT | O_RDWR | O_EXCL, S_IRUSR | S_IWUSR);
 	if (fd == -1) {
